float_update

+++ python_introduction/finance_calculator.py
@@ -1,5 +1,5 @@
-income = int (input ("Enter your monthly income: "))
-expense = int (input ("Enter your total monthly expense: "))
+income = float (input ("Enter your monthly income: "))
+expense = float (input ("Enter your total monthly expense: "))
 saving = income - expense
 Annual_saving = saving*12+(saving*12*0.05)
 Result = "Enter your monthly income: " + str(income) + "\n" +"Enter your total monthly expense: " + str (expense) + "\n" + "Your monthly savings are " + str (saving) +"\n" + "Projected savings after one year, with interest, is: " + str (Annual_saving)
